initial project

+++ storage-module/db/src/main/java/tipitapi/drawmytodayimprovement/util/PageableUtils.java
@@ -6,12 +6,12 @@ import tipitapi.drawmytodayimprovement.dto.PageResponse;
 import tipitapi.drawmytodayimprovement.dto.PageableRequest;
 
 public class PageableUtils {
-    public static Pageable toPageable(PageableRequest pageableRequest) {
-        return Pageable.ofSize(pageableRequest.size())
-                .withPage(pageableRequest.page());
-    }
+	public static Pageable toPageable(PageableRequest pageableRequest) {
+		return Pageable.ofSize(pageableRequest.size())
+				.withPage(pageableRequest.page());
+	}
 
-    public static <T> PageResponse<T> toPageResponse(Page<T> page) {
-        return new PageResponse<>(page.getNumber(), page.getSize(), (int) page.getTotalElements(), page.getContent());
-    }
+	public static <T> PageResponse<T> toPageResponse(Page<T> page) {
+		return new PageResponse<>(page.getNumber(), page.getSize(), page.getTotalElements(), page.getContent());
+	}
 }
